Change base class of PathField and move implementation
