+++ clickqt/widgets/filefield.py
@@ -1,5 +1,5 @@
 from PySide6.QtWidgets import QLineEdit, QInputDialog
-from clickqt.widgets.base_widget import PathField
+from clickqt.widgets.textfield import PathField
 from typing import Tuple, Any
 from clickqt.core.error import ClickQtError
 from click import Parameter
@@ -19,11 +19,11 @@ class FileFild(PathField):
             self.handleValid(True)
 
             def ret():
-                old_stdin = sys.stdin
                 user_input, ok = QInputDialog.getMultiLineText(self.widget, 'Stdin Input', self.label.text())
                 if not ok:
                     return (None, ClickQtError(ClickQtError.ErrorType.ABORTED_ERROR))
                 
+                old_stdin = sys.stdin
                 sys.stdin = BytesIO(user_input.encode(sys.stdin.encoding)) if "b" in self.param.type.mode else StringIO(user_input)
                 val = super(FileFild, self).getValue()
                 sys.stdin = old_stdin
@@ -32,6 +32,3 @@ class FileFild(PathField):
             return (ret, ClickQtError())
         else:
             return super().getValue()
-   
-    def getWidgetValue(self) -> str:
-        return self.widget.text()
